+++ src/component/AddLinkInput.js
@@ -1,0 +1,7 @@
+import styled from "../styles/AddLinkInput.css";
+
+function AddLinkInput(){
+
+}
+
+export default AddLinkInput
