--- src/pages/HomePage.js
@@ -1,10 +1,0 @@
-
-function HomePage(){
-    return(
-        <div>
-            홈페이지
-        </div>
-    )
-}
-
-export default HomePage;
