+++ src/pages/folderPage.js
@@ -4,8 +4,22 @@ import Banner from "../component/Banner";
 import CardList from "../component/CardList";
 import SearchInput from "../component/SearchInput";
 import styles from "../styles/FolderPage.module.css";
+import { getAllFolderData } from "../apis/folderApi";
+import { useEffect, useState } from "react";
 
-function FolderPage({ email, folder }) {
+function FolderPage({email}) {
+  const [folder, setFolder] = useState();
+
+  // const handleFolderLoad = async () => {
+  //   const { folder } = await getAllFolderData();
+  //   setFolder(folder);
+  //   console.log(folder);
+  // };
+
+  useEffect(()=>{
+    // handleFolderLoad()
+  },[])
+
   return (
     <>
       <Header email={email} />
