+++ src/styles/Header.module.css
@@ -1,8 +1,5 @@
 header {
   background-color: var(--gray-bg-color);
-  position: sticky;
-  top: 0;
-  z-index: 5;
   margin: 0 auto;
   display: flex;
   justify-content: center;
